removed .test, cleaned up yaaarghs
